updated code; everyones code into one file
Did a lot of changes, some things didn't work the way that they should have so I had to move some things around to try to get everything to fit together. Not perfect, but runs well on my end. Some formatting/ look issues, but values seem to be right and the map prints the way it should.